Holo Rotation,Cube Collection,Timer and Falling Off Conditions.
Holo now rotates to by arrow keys to the intended.
Cube Collapses when collided and adds the score to a Console 100 per each.
Timer is started at 120 Seond and keeps on Descending when it reaches 0 the scene is reloaded.
When the player falls off and reaches a certain collider, the scene is reloaded